Adds LoginCtrl controller and specs
Added sinonjs bower package so that we can stub and mock the
AuthService dependency (we havn't written this yet).
Needed to add 'sinon' to the list of included files and to .jshintrc
(so that undeclared variable is ignored).
Finally need to stub the promise that we expect the service to return.